Refactor A Bit More
